added xhci.c

+++ types.h
@@ -12,7 +12,7 @@ typedef uint8_t byte;
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint32_t u64;
+typedef uint64_t u64;
 
 #define ELFABI __attribute__((sysv_abi))
 #define SYSVABI __attribute__((sysv_abi))
